@@ -15,6 +15,6 @@ Gem::Specification.new do |s|
 
   s.files = Dir["{app,config,db,lib}/**/*"] + %w(CHANGELOG.md README.md)
 
-  s.add_dependency "rails", "~> 3.2.14"
+  s.add_dependency "rails", "~> 4.2.5"
   
 end
