@@ -16,11 +16,12 @@ module OpenProject::GwdgProjectGroups
         def project_settings_tabs_with_gwdg_project_groups
           tabs = project_settings_tabs_without_gwdg_project_groups
           if User.current.allowed_to?(:manage_project_groups, @project)
-            tabs.push({:name => 'project_groups',
-                       :action => :manage_project_groups,
-                       :partial => 'projects/settings/project_groups',
-                       :label => :label_group_plural})
-            #tabs.push({ name: 'project_groups', action: :manage_project_groups, partial: 'projects/settings/project_groups', label: :label_group_plural })
+            tabs.push({
+                        name: 'project_groups',
+                        action: :manage_project_groups,
+                        partial: 'projects/settings/project_groups',
+                        label: :label_group_plural
+                      })
           end
           tabs
         end
