@@ -1,7 +1,7 @@
 # -*- encoding : utf-8 -*-
 class CreateProjectGroupScopes < ActiveRecord::Migration
   def self.up
-    create_table :project_group_scopes, :id => true do |t| #XXX There should be no ID column but Postgres w/ ObjectDaddy don't like it
+    create_table :project_group_scopes do |t|
       t.column :project_id, :integer, :null => false
       t.column :project_group_id, :integer, :null => false
     end
